refactored the code & adapted to the new project structure
